requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.4.2
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
jinja2==3.1.2
python-multipart==0.0.6
werkzeug==2.3.7
psycopg2-binary==2.9.7
